+++ lib/oban.ex
@@ -148,7 +148,7 @@ defmodule Oban do
   use Oban.Testing, repo: MyApp.Repo
   ```
 
-  Now you can assert or refute that jobs have been enqueued within your tests:
+  Now you can assert, refute or list jobs that have been enqueued within your tests:
 
   ```elixir
   assert_enqueued worker: MyWorker, args: %{id: 1}
@@ -156,6 +156,10 @@ defmodule Oban do
   # or
 
   refute_enqueued queue: "special", args: %{id: 2}
+
+  # or
+
+  assert [%{args: %{"id" => 1}}] = all_enqueued worker: MyWorker
   ```
 
   See the `Oban.Testing` module for more details on making assertions.